+++ tests/providers/aws/services/lightsail/lightsail_instance_automated_snapshots/lightsail_instance_automated_snapshots_test.py
@@ -17,12 +17,15 @@ class Test_lightsail_instance_automated_snapshots:
         lightsail_client = MagicMock
         lightsail_client.instances = {}
 
-        with patch(
-            "prowler.providers.common.provider.Provider.get_global_provider",
-            return_value=set_mocked_aws_provider([AWS_REGION_US_EAST_1]),
-        ), patch(
-            "prowler.providers.aws.services.lightsail.lightsail_service.Lightsail",
-            new=lightsail_client,
+        with (
+            patch(
+                "prowler.providers.common.provider.Provider.get_global_provider",
+                return_value=set_mocked_aws_provider([AWS_REGION_US_EAST_1]),
+            ),
+            patch(
+                "prowler.providers.aws.services.lightsail.lightsail_service.Lightsail",
+                new=lightsail_client,
+            ),
         ):
             from prowler.providers.aws.services.lightsail.lightsail_instance_automated_snapshots.lightsail_instance_automated_snapshots import (
                 lightsail_instance_automated_snapshots,
@@ -66,12 +69,15 @@ class Test_lightsail_instance_automated_snapshots:
             )
         }
 
-        with patch(
-            "prowler.providers.common.provider.Provider.get_global_provider",
-            return_value=set_mocked_aws_provider([AWS_REGION_US_EAST_1]),
-        ), patch(
-            "prowler.providers.aws.services.lightsail.lightsail_service.Lightsail",
-            new=lightsail_client,
+        with (
+            patch(
+                "prowler.providers.common.provider.Provider.get_global_provider",
+                return_value=set_mocked_aws_provider([AWS_REGION_US_EAST_1]),
+            ),
+            patch(
+                "prowler.providers.aws.services.lightsail.lightsail_service.Lightsail",
+                new=lightsail_client,
+            ),
         ):
             from prowler.providers.aws.services.lightsail.lightsail_instance_automated_snapshots.lightsail_instance_automated_snapshots import (
                 lightsail_instance_automated_snapshots,
@@ -126,12 +132,15 @@ class Test_lightsail_instance_automated_snapshots:
             )
         }
 
-        with patch(
-            "prowler.providers.common.provider.Provider.get_global_provider",
-            return_value=set_mocked_aws_provider([AWS_REGION_US_EAST_1]),
-        ), patch(
-            "prowler.providers.aws.services.lightsail.lightsail_service.Lightsail",
-            new=lightsail_client,
+        with (
+            patch(
+                "prowler.providers.common.provider.Provider.get_global_provider",
+                return_value=set_mocked_aws_provider([AWS_REGION_US_EAST_1]),
+            ),
+            patch(
+                "prowler.providers.aws.services.lightsail.lightsail_service.Lightsail",
+                new=lightsail_client,
+            ),
         ):
             from prowler.providers.aws.services.lightsail.lightsail_instance_automated_snapshots.lightsail_instance_automated_snapshots import (
                 lightsail_instance_automated_snapshots,