+++ tests/providers/microsoft365/services/admincenter/admincenter_users_between_two_and_four_global_admins/admincenter_users_between_two_and_four_global_admins_test.py
@@ -13,12 +13,15 @@ class Test_admincenter_users_between_two_and_four_global_admins:
         admincenter_client.audited_tenant = "audited_tenant"
         admincenter_client.audited_domain = DOMAIN
 
-        with mock.patch(
-            "prowler.providers.common.provider.Provider.get_global_provider",
-            return_value=set_mocked_microsoft365_provider(),
-        ), mock.patch(
-            "prowler.providers.microsoft365.services.admincenter.admincenter_users_between_two_and_four_global_admins.admincenter_users_between_two_and_four_global_admins.admincenter_client",
-            new=admincenter_client,
+        with (
+            mock.patch(
+                "prowler.providers.common.provider.Provider.get_global_provider",
+                return_value=set_mocked_microsoft365_provider(),
+            ),
+            mock.patch(
+                "prowler.providers.microsoft365.services.admincenter.admincenter_users_between_two_and_four_global_admins.admincenter_users_between_two_and_four_global_admins.admincenter_client",
+                new=admincenter_client,
+            ),
         ):
             from prowler.providers.microsoft365.services.admincenter.admincenter_users_between_two_and_four_global_admins.admincenter_users_between_two_and_four_global_admins import (
                 admincenter_users_between_two_and_four_global_admins,
@@ -35,12 +38,15 @@ class Test_admincenter_users_between_two_and_four_global_admins:
         admincenter_client.audited_tenant = "audited_tenant"
         admincenter_client.audited_domain = DOMAIN
 
-        with mock.patch(
-            "prowler.providers.common.provider.Provider.get_global_provider",
-            return_value=set_mocked_microsoft365_provider(),
-        ), mock.patch(
-            "prowler.providers.microsoft365.services.admincenter.admincenter_users_between_two_and_four_global_admins.admincenter_users_between_two_and_four_global_admins.admincenter_client",
-            new=admincenter_client,
+        with (
+            mock.patch(
+                "prowler.providers.common.provider.Provider.get_global_provider",
+                return_value=set_mocked_microsoft365_provider(),
+            ),
+            mock.patch(
+                "prowler.providers.microsoft365.services.admincenter.admincenter_users_between_two_and_four_global_admins.admincenter_users_between_two_and_four_global_admins.admincenter_client",
+                new=admincenter_client,
+            ),
         ):
             from prowler.providers.microsoft365.services.admincenter.admincenter_service import (
                 DirectoryRole,
@@ -83,12 +89,15 @@ class Test_admincenter_users_between_two_and_four_global_admins:
         admincenter_client.audited_tenant = "audited_tenant"
         admincenter_client.audited_domain = DOMAIN
 
-        with mock.patch(
-            "prowler.providers.common.provider.Provider.get_global_provider",
-            return_value=set_mocked_microsoft365_provider(),
-        ), mock.patch(
-            "prowler.providers.microsoft365.services.admincenter.admincenter_users_between_two_and_four_global_admins.admincenter_users_between_two_and_four_global_admins.admincenter_client",
-            new=admincenter_client,
+        with (
+            mock.patch(
+                "prowler.providers.common.provider.Provider.get_global_provider",
+                return_value=set_mocked_microsoft365_provider(),
+            ),
+            mock.patch(
+                "prowler.providers.microsoft365.services.admincenter.admincenter_users_between_two_and_four_global_admins.admincenter_users_between_two_and_four_global_admins.admincenter_client",
+                new=admincenter_client,
+            ),
         ):
             from prowler.providers.microsoft365.services.admincenter.admincenter_service import (
                 DirectoryRole,
@@ -142,12 +151,15 @@ class Test_admincenter_users_between_two_and_four_global_admins:
         admincenter_client.audited_tenant = "audited_tenant"
         admincenter_client.audited_domain = DOMAIN
 
-        with mock.patch(
-            "prowler.providers.common.provider.Provider.get_global_provider",
-            return_value=set_mocked_microsoft365_provider(),
-        ), mock.patch(
-            "prowler.providers.microsoft365.services.admincenter.admincenter_users_between_two_and_four_global_admins.admincenter_users_between_two_and_four_global_admins.admincenter_client",
-            new=admincenter_client,
+        with (
+            mock.patch(
+                "prowler.providers.common.provider.Provider.get_global_provider",
+                return_value=set_mocked_microsoft365_provider(),
+            ),
+            mock.patch(
+                "prowler.providers.microsoft365.services.admincenter.admincenter_users_between_two_and_four_global_admins.admincenter_users_between_two_and_four_global_admins.admincenter_client",
+                new=admincenter_client,
+            ),
         ):
             from prowler.providers.microsoft365.services.admincenter.admincenter_service import (
                 DirectoryRole,